+++ rpc/core/pipe.go
@@ -28,6 +28,8 @@ type P2P interface {
 }
 
 //----------------------------------------------
+// These package level globals come with setters
+// that are expected to be called only once, on startup
 
 var (
 	// external, thread safe interfaces
@@ -41,10 +43,11 @@ var (
 	p2pSwitch      P2P
 
 	// objects
-	pubKey    crypto.PubKey
-	genDoc    *types.GenesisDoc // cache the genesis structure
-	addrBook  *p2p.AddrBook
-	txIndexer txindex.TxIndexer
+	pubKey           crypto.PubKey
+	genDoc           *types.GenesisDoc // cache the genesis structure
+	addrBook         *p2p.AddrBook
+	txIndexer        txindex.TxIndexer
+	consensusReactor *consensus.ConsensusReactor
 
 	logger log.Logger
 )
@@ -87,6 +90,10 @@ func SetProxyAppQuery(appConn proxy.AppConnQuery) {
 
 func SetTxIndexer(indexer txindex.TxIndexer) {
 	txIndexer = indexer
+}
+
+func SetConsensusReactor(conR *consensus.ConsensusReactor) {
+	consensusReactor = conR
 }
 
 func SetLogger(l log.Logger) {