+++ scripts/txs/random.sh
@@ -10,10 +10,10 @@ PORT=$2
 
 for i in `seq 1 $N`; do
 	# store key value pair
-	KEY="abcd$i"
-	VALUE="dcba$i"
-	echo "$KEY:$VALUE"
-	curl 127.0.0.1:$PORT/broadcast_tx_sync?tx=\"$(toHex $KEY=$VALUE)\"
+	KEY=$(head -c 10 /dev/urandom)
+	VALUE="$i"
+	echo $(toHex $KEY=$VALUE)
+	curl 127.0.0.1:$PORT/broadcast_tx_sync?tx=0x$(toHex $KEY=$VALUE)
 done