+++ types/canonical_json.go
@@ -31,6 +31,14 @@ type CanonicalJSONVote struct {
 	Type    byte                 `json:"type"`
 }
 
+type CanonicalJSONHeartbeat struct {
+	Height           int        `json:"height"`
+	Round            int        `json:"round"`
+	Sequence         int        `json:"sequence"`
+	ValidatorAddress data.Bytes `json:"validator_address"`
+	ValidatorIndex   int        `json:"validator_index"`
+}
+
 //------------------------------------
 // Messages including a "chain id" can only be applied to one chain, hence "Once"
 
@@ -42,6 +50,11 @@ type CanonicalJSONOnceProposal struct {
 type CanonicalJSONOnceVote struct {
 	ChainID string            `json:"chain_id"`
 	Vote    CanonicalJSONVote `json:"vote"`
+}
+
+type CanonicalJSONOnceHeartbeat struct {
+	ChainID   string                 `json:"chain_id"`
+	Heartbeat CanonicalJSONHeartbeat `json:"heartbeat"`
 }
 
 //-----------------------------------
@@ -77,5 +90,15 @@ func CanonicalVote(vote *Vote) CanonicalJSONVote {
 		vote.Height,
 		vote.Round,
 		vote.Type,
+	}
+}
+
+func CanonicalHeartbeat(heartbeat *Heartbeat) CanonicalJSONHeartbeat {
+	return CanonicalJSONHeartbeat{
+		heartbeat.Height,
+		heartbeat.Round,
+		heartbeat.Sequence,
+		heartbeat.ValidatorAddress,
+		heartbeat.ValidatorIndex,
 	}
 }